requirements.txt
numpy
scipy
scikit-learn
matplotlib
torch
torchvision
umap-learn
tqdm